+++ Logo_Turtle/movimento.h
@@ -1,0 +1,6 @@
+#ifndef MOVIMENTO_H_INCLUDED
+#define MOVIMENTO_H_INCLUDED
+
+int movimenta_tartaruga(char matriz[largura][altura], entr quantidade[20]);
+
+#endif // MOVIMENTO_H_INCLUDED
