+++ Logo_Turtle/registro.h
@@ -1,10 +1,13 @@
 #ifndef REGISTRO_H_INCLUDED
 #define REGISTRO_H_INCLUDED
 ///def
-#define largura 50 //optei por deixar as duas variaveis simetricas
+
+#define largura 50
 #define altura 100
+
+
 ///signa
-void salvar_matriz(int matriz[largura][altura]);
+void salvar_matriz(char matriz[largura][altura]);
 void criar_arquivo();
 
 #endif // REGISTRO_H_INCLUDED
